refactor: styles and structure refactoring
