+++ src/features/Main/SearchResults/SearchResults.jsx
@@ -2,7 +2,7 @@ import styles from './SearchResults.module.scss';
 
 const SearchResults = () => (
   <div className={styles.search_results_wrapper}>
-    <b>39</b> movies found
+    <span className={styles.search_results_span}>39</span> movies found
   </div>
 );
