+++ src/features/Main/SearchResults/SearchResults.module.scss
@@ -3,3 +3,7 @@
   max-width: 1080px;
   margin-bottom: 20px;
 }
+
+.search_results_span {
+  font-weight: bold;
+}
